Classmind-v2/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
aiosmtplib>=3.0.0
google-auth>=2.29.0
requests>=2.31.0

